refactor: split data source options resolver
