+++ crates/sail-common-datafusion/src/datasource.rs
@@ -11,7 +11,6 @@ use datafusion::physical_plan::ExecutionPlan;
 use datafusion_common::{plan_err, Constraints, DFSchema, Result};
 use datafusion_expr::expr::Sort;
 use datafusion_expr::Expr;
-use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Eq, PartialEq, Hash, PartialOrd)]
 pub enum SinkMode {
@@ -141,14 +140,4 @@ pub fn get_partition_columns_and_file_schema(
         .collect::<Vec<_>>();
     let file_schema = Schema::new(file_schema_fields);
     Ok((partition_columns, file_schema))
-}
-
-/// Options that control the behavior of Delta Lake tables.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct TableDeltaOptions {
-    pub replace_where: Option<String>,
-    pub merge_schema: bool,
-    pub overwrite_schema: bool,
-    pub target_file_size: usize,
-    pub write_batch_size: usize,
 }
