+++ crates/sail-data-source/src/formats/delta.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 
@@ -17,11 +18,12 @@ use sail_delta_lake::delta_format::DeltaDataSink;
 use sail_delta_lake::operations::write::execution::{
     prepare_predicate_actions_physical, WriterStatsConfig,
 };
+use sail_delta_lake::options::TableDeltaOptions;
 use sail_delta_lake::table::open_table_with_object_store;
 use url::Url;
 use uuid::Uuid;
 
-use crate::options::DataSourceOptionsResolver;
+use crate::options::{load_default_options, load_options, DeltaReadOptions, DeltaWriteOptions};
 
 #[derive(Debug, Default)]
 pub struct DeltaTableFormat;
@@ -47,7 +49,8 @@ impl TableFormat for DeltaTableFormat {
             options,
         } = info;
         let table_url = Self::parse_table_url(ctx, paths).await?;
-        create_delta_provider(ctx, table_url, schema, &options).await
+        let options = resolve_delta_read_options(options)?;
+        create_delta_provider(ctx, table_url, schema, options).await
     }
 
     async fn create_writer(
@@ -70,8 +73,7 @@ impl TableFormat for DeltaTableFormat {
         }
 
         let table_url = Self::parse_table_url(ctx, vec![path]).await?;
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let delta_options = resolver.resolve_delta_write_options(options)?;
+        let delta_options = resolve_delta_write_options(options)?;
 
         let mut initial_actions: Vec<Action> = Vec::new();
         let mut operation: Option<DeltaOperation> = None;
@@ -267,4 +269,51 @@ impl DeltaTableFormat {
             _ => plan_err!("expected a single path for Delta table sink: {paths:?}"),
         }
     }
+}
+
+fn apply_delta_read_options(from: DeltaReadOptions, to: &mut TableDeltaOptions) -> Result<()> {
+    // TODO: implement read options
+    let _ = (from, to);
+    Ok(())
+}
+
+fn apply_delta_write_options(from: DeltaWriteOptions, to: &mut TableDeltaOptions) -> Result<()> {
+    if let Some(replace_where) = from.replace_where {
+        to.replace_where = Some(replace_where);
+    }
+    if let Some(merge_schema) = from.merge_schema {
+        to.merge_schema = merge_schema;
+    }
+    if let Some(overwrite_schema) = from.overwrite_schema {
+        to.overwrite_schema = overwrite_schema;
+    }
+    if let Some(target_file_size) = from.target_file_size {
+        to.target_file_size = target_file_size;
+    }
+    if let Some(write_batch_size) = from.write_batch_size {
+        to.write_batch_size = write_batch_size;
+    }
+    Ok(())
+}
+
+pub fn resolve_delta_read_options(
+    options: Vec<HashMap<String, String>>,
+) -> Result<TableDeltaOptions> {
+    let mut delta_options = TableDeltaOptions::default();
+    apply_delta_read_options(load_default_options()?, &mut delta_options)?;
+    for opt in options {
+        apply_delta_read_options(load_options(opt)?, &mut delta_options)?;
+    }
+    Ok(delta_options)
+}
+
+pub fn resolve_delta_write_options(
+    options: Vec<HashMap<String, String>>,
+) -> Result<TableDeltaOptions> {
+    let mut delta_options = TableDeltaOptions::default();
+    apply_delta_write_options(load_default_options()?, &mut delta_options)?;
+    for opt in options {
+        apply_delta_write_options(load_options(opt)?, &mut delta_options)?;
+    }
+    Ok(delta_options)
 }
