+++ crates/sail-data-source/src/formats/listing.rs
@@ -6,11 +6,6 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::catalog::{Session, TableProvider};
-use datafusion::datasource::file_format::arrow::ArrowFormat;
-use datafusion::datasource::file_format::avro::AvroFormat;
-use datafusion::datasource::file_format::csv::CsvFormat;
-use datafusion::datasource::file_format::json::JsonFormat;
-use datafusion::datasource::file_format::parquet::ParquetFormat;
 use datafusion::datasource::file_format::FileFormat;
 use datafusion::datasource::listing::{ListingOptions, ListingTable, ListingTableConfig};
 use datafusion::datasource::physical_plan::FileSinkConfig;
@@ -23,8 +18,6 @@ use sail_common_datafusion::datasource::{
     get_partition_columns_and_file_schema, SinkInfo, SourceInfo, TableFormat,
 };
 
-use crate::formats::text::file_format::TextFileFormat;
-use crate::options::DataSourceOptionsResolver;
 use crate::utils::split_parquet_compression_string;
 
 // TODO: support global configuration to ignore file extension (by setting it to empty)
@@ -231,208 +224,5 @@ impl<T: ListingFormat> TableFormat for ListingTableFormat<T> {
         format
             .create_writer_physical_plan(input, ctx, conf, sort_order)
             .await
-    }
-}
-
-// Arrow
-pub(crate) type ArrowTableFormat = ListingTableFormat<ArrowListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct ArrowListingFormat;
-
-impl ListingFormat for ArrowListingFormat {
-    fn name(&self) -> &'static str {
-        "arrow"
-    }
-
-    fn create_read_format(
-        &self,
-        _ctx: &dyn Session,
-        _options: Vec<HashMap<String, String>>,
-        _compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        Ok(Arc::new(ArrowFormat))
-    }
-
-    fn create_write_format(
-        &self,
-        _ctx: &dyn Session,
-        _options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        Ok((Arc::new(ArrowFormat), None))
-    }
-}
-
-// Avro
-pub(crate) type AvroTableFormat = ListingTableFormat<AvroListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct AvroListingFormat;
-
-impl ListingFormat for AvroListingFormat {
-    fn name(&self) -> &'static str {
-        "avro"
-    }
-
-    fn create_read_format(
-        &self,
-        _ctx: &dyn Session,
-        _options: Vec<HashMap<String, String>>,
-        _compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        Ok(Arc::new(AvroFormat))
-    }
-
-    fn create_write_format(
-        &self,
-        _ctx: &dyn Session,
-        _options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        Ok((Arc::new(AvroFormat), None))
-    }
-}
-
-// Csv
-pub(crate) type CsvTableFormat = ListingTableFormat<CsvListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct CsvListingFormat;
-
-impl ListingFormat for CsvListingFormat {
-    fn name(&self) -> &'static str {
-        "csv"
-    }
-
-    fn create_read_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-        compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let mut options = resolver.resolve_csv_read_options(options)?;
-        if let Some(compression) = compression {
-            options.compression = compression;
-        }
-        Ok(Arc::new(CsvFormat::default().with_options(options)))
-    }
-
-    fn create_write_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let options = resolver.resolve_csv_write_options(options)?;
-        Ok((Arc::new(CsvFormat::default().with_options(options)), None))
-    }
-}
-
-// Json
-pub(crate) type JsonTableFormat = ListingTableFormat<JsonListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct JsonListingFormat;
-
-impl ListingFormat for JsonListingFormat {
-    fn name(&self) -> &'static str {
-        "json"
-    }
-
-    fn create_read_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-        compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let mut options = resolver.resolve_json_read_options(options)?;
-        if let Some(compression) = compression {
-            options.compression = compression;
-        }
-        Ok(Arc::new(JsonFormat::default().with_options(options)))
-    }
-
-    fn create_write_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let options = resolver.resolve_json_write_options(options)?;
-        Ok((Arc::new(JsonFormat::default().with_options(options)), None))
-    }
-}
-
-// Parquet
-pub(crate) type ParquetTableFormat = ListingTableFormat<ParquetListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct ParquetListingFormat;
-
-impl ListingFormat for ParquetListingFormat {
-    fn name(&self) -> &'static str {
-        "parquet"
-    }
-
-    fn create_read_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-        _compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let options = resolver.resolve_parquet_read_options(options)?;
-        Ok(Arc::new(ParquetFormat::default().with_options(options)))
-    }
-
-    fn create_write_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let options = resolver.resolve_parquet_write_options(options)?;
-        let compression = options.global.compression.clone();
-        Ok((
-            Arc::new(ParquetFormat::default().with_options(options)),
-            compression,
-        ))
-    }
-}
-
-// Text
-pub(crate) type TextTableFormat = ListingTableFormat<TextListingFormat>;
-
-#[derive(Debug, Default)]
-pub(crate) struct TextListingFormat;
-
-impl ListingFormat for TextListingFormat {
-    fn name(&self) -> &'static str {
-        "text"
-    }
-
-    fn create_read_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-        compression: Option<CompressionTypeVariant>,
-    ) -> Result<Arc<dyn FileFormat>> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let mut options = resolver.resolve_text_read_options(options)?;
-        if let Some(compression) = compression {
-            options.compression = compression;
-        }
-        Ok(Arc::new(TextFileFormat::new(options)))
-    }
-
-    fn create_write_format(
-        &self,
-        ctx: &dyn Session,
-        options: Vec<HashMap<String, String>>,
-    ) -> Result<(Arc<dyn FileFormat>, Option<String>)> {
-        let resolver = DataSourceOptionsResolver::new(ctx);
-        let options = resolver.resolve_text_write_options(options)?;
-        Ok((Arc::new(TextFileFormat::new(options)), None))
     }
 }
