+++ crates/sail-data-source/src/formats/mod.rs
@@ -1,3 +1,8 @@
+pub mod arrow;
+pub mod avro;
+pub mod csv;
 pub mod delta;
+pub mod json;
 pub mod listing;
+pub mod parquet;
 pub mod text;
