+++ crates/sail-delta-lake/Cargo.toml
@@ -4,9 +4,6 @@ version = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-sail-common = { path = "../sail-common" }
-sail-common-datafusion = { path = "../sail-common-datafusion" }
-
 # Delta Lake
 deltalake = { workspace = true }
 delta_kernel = { workspace = true }
