+++ crates/sail-delta-lake/src/lib.rs
@@ -2,6 +2,7 @@ pub mod delta_datafusion;
 pub mod delta_format;
 mod kernel;
 pub mod operations;
+pub mod options;
 pub mod table;
 
 pub use table::create_delta_provider;
